requirements/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Image processing and OpenGL
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.25.2

# Validation and parsing
# TODO: Add a GLSL parser or custom implementation here
lark==1.1.8

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
PyJWT==2.8.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# WebSocket support
websockets==12.0

# HTTP client
httpx==0.25.2

# JSON processing
orjson==3.9.10

# Date and time
python-dateutil==2.8.2

# Configuration
dynaconf==3.2.4 